requirements.txt
langchain==0.3.27
langchain-openai==0.3.32
django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
python-dotenv==1.1.1

